pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neofs-python-sdk"
version = "0.1.0"
description = "Pure Python NeoFS SDK and CLI"
authors = [
    { name = "NeoFS Python Developer" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "protobuf>=4.25.0",
    "neo-mamba>=0.8.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "base58>=2.1.1",
]

[project.scripts]
neofs = "neofs.cli:app"
